Kill menuSessionId variable
