+++ src/main/java/application/DebuggerController.java
@@ -1,5 +1,6 @@
 package application;
 
+import annotations.AppInstall;
 import annotations.AppInstallFromSession;
 import annotations.UserLock;
 import annotations.UserRestore;
@@ -76,14 +77,13 @@ public class DebuggerController extends AbstractBaseController {
     @ApiOperation(value = "Get content needed for the menu debugger")
     @RequestMapping(value = Constants.URL_DEBUGGER_MENU_CONTENT, method = RequestMethod.POST)
     @UserRestore
-    @AppInstallFromSession
+    @AppInstall
     public MenuDebuggerContentResponseBean menuDebuggerContent(
-            @RequestBody MenuDebuggerRequestBean debuggerMenuRequest,
+            @RequestBody SessionNavigationBean debuggerMenuRequest,
             @CookieValue(Constants.POSTGRES_DJANGO_SESSION_ID) String authToken) throws Exception {
 
-        MenuSession menuSession = getMenuSession(
-                debuggerMenuRequest.getMenuSessionId()
-        );
+        MenuSession menuSession = getMenuSessionFromBean(debuggerMenuRequest);
+        advanceSessionWithSelections(menuSession, debuggerMenuRequest.getSelections());
 
         return new MenuDebuggerContentResponseBean(
                 menuSession.getAppId(),
@@ -91,6 +91,34 @@ public class DebuggerController extends AbstractBaseController {
                 menuSession.getSessionWrapper().getEvaluationContext().getInstanceIds(),
                 fetchRecentMenuXPathQueries(debuggerMenuRequest.getDomain(), debuggerMenuRequest.getUsername())
         );
+    }
+
+    @ApiOperation(value = "Evaluate the given XPath under the current context")
+    @RequestMapping(value = Constants.URL_EVALUATE_MENU_XPATH, method = RequestMethod.POST)
+    @ResponseBody
+    @UserLock
+    @UserRestore
+    @AppInstall
+    public EvaluateXPathResponseBean menuEvaluateXpath(@RequestBody EvaluateXPathMenuRequestBean evaluateXPathRequestBean,
+                                                       @CookieValue(Constants.POSTGRES_DJANGO_SESSION_ID) String authToken) throws Exception {
+        MenuSession menuSession = getMenuSessionFromBean(evaluateXPathRequestBean);
+        advanceSessionWithSelections(menuSession, evaluateXPathRequestBean.getSelections());
+
+
+        EvaluateXPathResponseBean evaluateXPathResponseBean = new EvaluateXPathResponseBean(
+                menuSession.getSessionWrapper().getEvaluationContext(),
+                evaluateXPathRequestBean.getXpath()
+        );
+
+        cacheMenuXPathQuery(
+                evaluateXPathRequestBean.getDomain(),
+                evaluateXPathRequestBean.getUsername(),
+                evaluateXPathRequestBean.getXpath(),
+                evaluateXPathResponseBean.getOutput(),
+                evaluateXPathResponseBean.getStatus()
+        );
+
+        return evaluateXPathResponseBean;
     }
 
     @ApiOperation(value = "Evaluate the given XPath under the current context")
@@ -108,34 +136,6 @@ public class DebuggerController extends AbstractBaseController {
         );
 
         cacheFormXPathQuery(
-                evaluateXPathRequestBean.getDomain(),
-                evaluateXPathRequestBean.getUsername(),
-                evaluateXPathRequestBean.getXpath(),
-                evaluateXPathResponseBean.getOutput(),
-                evaluateXPathResponseBean.getStatus()
-        );
-
-        return evaluateXPathResponseBean;
-    }
-
-    @ApiOperation(value = "Evaluate the given XPath under the current context")
-    @RequestMapping(value = Constants.URL_EVALUATE_MENU_XPATH, method = RequestMethod.POST)
-    @ResponseBody
-    @UserLock
-    @UserRestore
-    @AppInstallFromSession
-    public EvaluateXPathResponseBean menuEvaluateXpath(@RequestBody EvaluateXPathMenuRequestBean evaluateXPathRequestBean,
-                                                       @CookieValue(Constants.POSTGRES_DJANGO_SESSION_ID) String authToken) throws Exception {
-        MenuSession menuSession = getMenuSession(
-                evaluateXPathRequestBean.getMenuSessionId()
-        );
-
-        EvaluateXPathResponseBean evaluateXPathResponseBean = new EvaluateXPathResponseBean(
-                menuSession.getSessionWrapper().getEvaluationContext(),
-                evaluateXPathRequestBean.getXpath()
-        );
-
-        cacheMenuXPathQuery(
                 evaluateXPathRequestBean.getDomain(),
                 evaluateXPathRequestBean.getUsername(),
                 evaluateXPathRequestBean.getXpath(),
