+++ src/main/java/beans/EvaluateXPathMenuRequestBean.java
@@ -9,7 +9,7 @@ import session.MenuSession;
  * Bean used for evaluating xpath against a given menu session.
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class EvaluateXPathMenuRequestBean extends  InstallFromSessionRequestBean {
+public class EvaluateXPathMenuRequestBean extends SessionNavigationBean {
     private String xpath;
 
     // default constructor for Jackson
@@ -25,16 +25,6 @@ public class EvaluateXPathMenuRequestBean extends  InstallFromSessionRequestBean
     }
 
     public String toString(){
-        return "EvaluateXPathRequestBean [xpath: " + xpath + ", menuSessionId: " + menuSessionId + "]";
-    }
-
-    @JsonSetter(value = "session_id")
-    public void setMenuSessionId(String menuSessionId) {
-        this.menuSessionId = menuSessionId;
-    }
-
-    @JsonGetter(value = "session_id")
-    public String getMenuSessionId() {
-        return menuSessionId;
+        return "EvaluateXPathRequestBean [xpath: " + xpath + "]";
     }
 }
